HW-6 fix: jdbc/jpa packaging
