+++ scrapper/src/main/java/edu/java/configuration/JdbcAccessConfiguration.java
@@ -3,9 +3,9 @@ package edu.java.configuration;
 import edu.java.dao.ChatDao;
 import edu.java.dao.ChatLinkDao;
 import edu.java.dao.LinkDao;
-import edu.java.jdbc.JdbcChatLinkService;
-import edu.java.jdbc.JdbcChatService;
-import edu.java.jdbc.JdbcLinkService;
+import edu.java.jdbc.service.JdbcChatLinkService;
+import edu.java.jdbc.service.JdbcChatService;
+import edu.java.jdbc.service.JdbcLinkService;
 import edu.java.service.ChatLinkService;
 import edu.java.service.ChatService;
 import edu.java.service.LinkService;
