+++ scrapper/src/main/java/edu/java/jdbc/dao/JdbcLinkDao.java
@@ -1,4 +1,4 @@
-package edu.java.jdbc;
+package edu.java.jdbc.dao;
 
 import edu.java.dao.LinkDao;
 import edu.java.dto.LinkDTO;
