+++ scrapper/src/main/java/edu/java/jpa/repository/ChatRepository.java
@@ -1,4 +1,4 @@
-package edu.java.jpa;
+package edu.java.jpa.repository;
 
 import edu.java.domain.Chat;
 import org.springframework.data.jpa.repository.JpaRepository;
