+++ scrapper/src/test/java/edu/java/jpa/service/JpaLinkServiceTest.java
@@ -1,8 +1,9 @@
-package edu.java.jpa;
+package edu.java.jpa.service;
 
 import edu.java.dto.LinkDTO;
 import edu.java.exception.LinkAlreadyAddedException;
 import edu.java.exception.LinkNotFoundException;
+import edu.java.jpa.repository.LinkRepository;
 import edu.java.scrapper.IntegrationTest;
 import java.time.LocalDateTime;
 import java.util.Collection;
